move regress out to QC
